+++ src/redtail-authentication/index.ts
@@ -3,4 +3,6 @@ export { RedtailAuthenticationController } from './redtail-authentication.contro
 export { RedtailAuthenticationMiddleware } from './redtail-authentication.middleware';
 export { RedtailAuthenticationModule } from './redtail-authentication.module';
 export { RedtailAuthenticationService } from './redtail-authentication.service';
+export { RedtailHttpInterceptor } from './redtail-http-interceptor.interceptor';
 export { UtilsService } from './utils.service';
+